fix(styling): Share a global Emotion instance between CK instances (#3152)
If Canvas Kit Styling is duplicated by bundling a copy, a new Emotion cache instance is created. The fix is to share an Emotion cache via the window object (if available). This will both decrease duplicated styles being injected into the page as well as fix style merging issues when a bundle is loaded after another bundle already loaded and rendered.

Fixes: #3151

[category:Infrastructure]

Release Note:
This change shares an Emotion cache instance on the window object. This will fix style merge issues in edge cases where a bundle's base styles override override styles of another bundle. It is unlikely that anyone relies on this broken functionality, but if anyone does, styles may be different. This change will only affect style merging from this release on and will not affect any previous version, so it is an opt-in.