docs: Fix Firefox hard-fail on markdown-to-jsx bug (#3153)
The `markdown-to-jsx` seems to have a bug in Firefox that causes improperly formatted HTML. This was causing an edge case that isn't supposed to be possible in our documentation system. Instead of Firefox erroring on proper parsing, we instead guard an access. Firefox will now show improperly formatted content rather than erroring out with no content. We will address the Firefox bug with our Vite upgrade where we drop `markdown-to-jsx` which should fix the issue.

[category:Documentation]