fix: Add minWidth and minHeight 1px to accessibleHide (#3206)
Fixes: [#3187 ](https://github.com/Workday/canvas-kit/issues/3187)

FormFieldLabel contains a minWidth of 180px which is preventing the width of 1px from accessibleHide to take effect. 
This fix will add a minWidth and minHeight to accessibleHide to override any potential minWidth or minHeight from parent component.

[category:Components]

Release Note:
isHidden prop in `FormField.Label` will now update the minWidth and minHeight to 1px.

Co-authored-by: Tu Pham <tu.pham@workday.com>