feat!: Promote new Dub logos and add codemod to handle change (#3201)
Fixes: #3150  

This PR removes the old Dub Logos in Main and promotes the new logos that were previously in our Preview package to Main as part of the brand refresh. It also adds a codemod to update the imports and update the logo names.

[category:Branding]

Release Note:
We've removed the old Dub Logos in Main and promoted the new logos that were previously in our Preview package to Main as part of the brand refresh. We've also added a codemod to update the import statements as well as the name of the logos used to the new names. 

| Old Logo Name | New Logo Name |
| --------------- | ----------------- | 
| dubLogoBlue | dubLogoPrimary | 
| dubLogoWhite | dubLogoReversed | 
| wdayLogoBlue | wdayLogoPrimary | 
| wdayLogoWhite | wdayLogoReversed |

### BREAKING CHANGES
There is a codemod that should catch a lot of instances but, if there is an instance that is not caught, users will have to update those names manually. See above table for name references.