feat!: Promote Expandable and add codemod to handle change (#3217)
Fixes: #3106  

This PR promotes the `Expandable` component from Labs to Main. It also adds a codemod to update the imports.

[category:Components]

Release Note:
Promotion of the `Expandable` component from Labs to Main.

### BREAKING CHANGES
The `Expandable` component has been promoted from `Labs` to `Main`. There is a codemod to handle the update of the imports.