fix: Update InformationHighlight.Body to use div (#3222)
The `InformationHighlight.Body` subcomponent was previously a `p` element, which is _fine_ on its own, but it makes things unintuitive when you need to add multiple blocks of content inside the `Body`. Nesting `p` elements or adding list elements (`ul` or `ol`) is not allowed. To solve this, we're making `Body` a generic `div` element. This is consistent with our other `Body` subcomponents such as `Modal.Body`.

Fixes: #3218

[category:Components]

Release Note:
There are no visual changes because of this update. However, if you were relying on the `p` tag as an identifier for testing or other purposes, you'll need to adjust your implementation, as the underlying element is now a `div`. Additionally, if you'd like to keep a `p` tag around the `Body` text, you'll need to add it manually. Wrapping text in a `p` is not an accessibility requirement, but it can be helpful when you have multiple, separate content blocks.

```jsx
// before
<InformationHighlight.Body>
  Body text goes here.
  <ul>
    <li>Point One</li>
    <li>Point Two</li>
    <li>Point Three</li>
  </ul>
</InformationHighlight.Body>

// after
<InformationHighlight.Body>
  <p style={{ margin: 0 }}>Body text goes here.</p>
  <ul>
    <li>Point One</li>
    <li>Point Two</li>
    <li>Point Three</li>
  </ul>
</InformationHighlight.Body>
```