fix: PopupStack.contains recursively determines ownership (#3262)
`useAlwaysCloseOnOutsideClick` doesn't work correctly because `PopupStack.contains` doesn't recursively check popup ownership. It only checks the first level. If we have more than two layers of popups that use `useAlwaysCloseOnOutsideClick`, the first popup will close when the third popup is clicked on which will prevent normal operation of the third popup. In a hierarchical menu, it means you could not select an option from the third menu.

[category:Components]

Release Note:
This should not negatively impact anyone. If your application relies on the broken functionality of `useAlwaysCloseOnOutsideClick`, your application may behave differently. If you have a workaround, the workaround will no longer be necessary.