feat: Use css transitions for sidepanel instead of animation (#3258)
Use css `transition` instead of animation in SidePanel (Preview).

[category:Components]

Release Note:
When we supported IE 11 we needed to use animation to support animation events. Since we dropped support for IE11, we can now use css transitions to animate the width of the SidePanel when it's expanding or collapsing.  

As part of this change, we also removed the `onAnimationStart` prop. Since moving to transitions, React doesn't support `onTransitionStart`, therefore the prop is no longer needed.

Co-authored-by: manuel.carrera <manuel.carrera@workday.com>
Co-authored-by: @RayRedGoose <48605821+RayRedGoose@users.noreply.github.com>