ci: Pad canary build ids with zeros
We have issues with pre-ids in our canary builds where package managers sort alphabetically. This is an issue when we go from `999` to `1000`. When sorting alphabetically, "1000" is before "999" instead of after. If we prefix with a zero, we can handle up to 9999 before we have this issue. `0999` is properly sorted before `1000`