feat: Update Color Picker to use new color palette from brand (#3307)
Fixes: #3273

We've updated `ColorPicker` components in our Main and Preview packages to use our tokens and styling utilities. The React interface has not changed, but CSS variables are now used for dynamic properties.

[category:Components]

Release Note:
We've updated `ColorPicker` components in Main and Preview packages to use our new tokens and styling utilities. The React interface has not changed, but CSS variables are now used for dynamic properties.

Co-authored-by: @mannycarrera4 <mannycarrera4@users.noreply.github.com>