fix(tabs): Apply fade more appropriately for mobile (#3380)
Update the logic around the Tab list fade effect when scrolling with a touch device. The original logic applied a fade effect only when touch-scrolling and only in the direction of the scroll. The original intent was to apply a fade to indicate where content was being overflowed. The update adds scroll detection logic and applies a fade only to directions that can be scrolled to.

Fixes: #3377

[category:Components]