fix: Use dvh in Popup.Card height calculation (#3419)
Fixes https://github.com/Workday/canvas-kit/issues/3384
Use `dvh` units instead of `vh` when calculating the height of the Popup.Card.

`dvh` factors in mobile-specific features like the status and address bars when making calculations, unlike `vh`. It is supported on all major browsers at this time according to [caniuse](https://caniuse.com/mdn-css_types_length_viewport_percentage_units_dynamic)

When using `vh`, the browser will not account for these elements and so in certain cases the content may be cut off as shown in the issue. Normally this will not be a problem, but the `maxHeight` will fail to prevent overflow if the height attribute is overridden by the consumer.

[category:Components]