fix(collection): Fix `inert` prop in React 19 (#3433)
Fix dev warning for React 19 and fix `inert` in React 19. In React <19, adding `inert=''` added the attribute and the correct behavior. In React 19, `inert=''` does not add the attribute. `'true'` works in all versions of React, but triggers a warning in React 19, so we detect React version and set the value differently to ensure correct runtime behavior and avoid any React warnings.

[category:Components]