feat!: Remove deprecated orange buttons (#3439)
Fixes: https://github.com/Workday/canvas-kit/issues/3395

Long overdue, but a sign of moving forward, we are removing our `DeprecatedButton`. Our design system supported this for quite some time, but with the advancement in theming, our components API, and our flexibility, it is time for us to finally remove this component. This component has served as a reminder of how far we've come and it part we're thankful for this change.

[category:Components]

### BREAKING CHANGES
Al though unlikely, but not impossible, if you were using `DeprecatedButton` please use our flexible and themable buttons like `PrimaryButton`, `ScondaryButon` or `TertiaryButton`.

Co-authored-by: manuel.carrera <manuel.carrera@workday.com>