fix: Update popupstack to account for scoped theming (#3635)
Fixes: https://github.com/Workday/canvas-kit/issues/3636
We initially removed className and style forwarding to popups and modals which helped global theming but broke scoped theming. This change allows global theming if no theme is provided, if a theme is provided via the canvas provider, **only** the values defined are forward to popup and modals.

[category:Components]

Release Note:
[This change](https://github.com/Workday/canvas-kit/pull/3572) completely removed theme forwarding to popups and modals. However, what we really want is to **only** forward theme and the tokens defined by theme if they are provided, otherwise allow global tokens to trickle through to popups and modals.

Previously, the `usePopupStack` hook created a CSS class name that was passed to our Popups. We
attached those theme styles to that class name. This allowed the theme to be available in our
Popups. But it also created a cascade barrier that blocked the global theme from being applied to
our Popup components. Because we now use global CSS variables, we no longer need this class name to
provide the global theme to Popups. But we have to remove this generated class name to allow the
global theme to be applied to Popups.

If you want to have scoped theming where a part of your application needs different theming, you can
still do this via the `theme` prop.

> **Note:** Only the properties of the theme object that are changed will be forward to popups and
> modals. IE, if you change theme.palette.primary.main, only those tokens will change for popups and
> modals.

Co-authored-by: manuel.carrera <manuel.carrera@workday.com>